requirements_llamacpp.txt
boto3==1.34.162
gritlm==1.0.2
llama-cpp-python==0.3.5
